[#37] [Experimental] More accurate clock-total when merging ordered disjoint pstats
This was motivated by a PR from @ivarref:
https://github.com/ptaoussanis/tufte/pull/37/commits/09db6f03fd7cc89bc47b324e267461497981fd7f

In particular, we make the observation that when merges are serial (common case),
we'll be merging PStats in increasing order of t1 (completion time). I.e. clock
intervals (time spans) will actually already be ~pre-sorted by t1.

So we should be able to avoid the storage, cost, and complexity of a  non-optional
O(nlogn) pre-sort.

What if merges are done against unordered PStats? We can detect this case easily
enough and flag the clock data as approximate. This way users can decide for
themselves whether they want to pay for a pre-sort, or to consider a serialization
mechanism (e.g. agent).