[#52 #35] Introduce time-spans for accurate clock total under all conditions (@ivarref)
Previously, when merging `PStats` - it was possible for the reported Clock total time
to be significantly too low.

This'd happen in situations like the following, with `profiled` forms being
accumulated in order of *completion* time (1, 2, 3):

   <------------------------------------------------> ; Time
1:  [profiled1   ]
2:                               [profiled2  ]
3:          [profiled3                           ]

I.e.
  - `profiled1` starts
  - `profiled3` starts
  - `profiled1` completes (1), gets accumulated
  - `profiled2` starts
  - `profiled2` completes (2), gets accumulated (causing bump of merged `t1`)
  - `profiled3` completes (3), gets accumulated but at this point all we have is:
     - [acc_t0=p1_t0, acc_t1=p2_t1]
     - [       p3_t0,        p3_t1]
     - And there's actually no way to losslessly merge our clock any more.

There's no simple solution to this problem. Originally hoped that we might
be able to get away with ensuring that spans to merge have a
monotonically-increasing right interval (which is inexpensively possible).

Unfortunately turns out that what we need is really monotonically-increasing
*left* intervals. That's more difficult to get, and basically requires sorting
intervals before merging- which is O(NlogN).

The solution as of this commit was provided by @ivarref and should allow an
accurate (lossless) clock merge. Some amount of lossyness will occur only
in rare cases when a very large number of `PStats` are being merged,
triggering compaction.

Otherwise we're basically getting clock accuracy here in exchange for
O(NlogN) cost on `PStats` deref | N is number of merged `PStats`. This is
probably a decent exchange since folks aren't typically merging a large number
of PStats (i.e. N will be small in most cases), and since the merge cost
can mostly be paid on a background thread.

Anyway this mostly affects folks using something like the StatsAccumulator.

Big thanks to Ivar for reporting this, and for the provided solution.