[mod] Rename format columns: `:n-calls` -> `:n`, `:total` -> `:sum`
Previously the `columns` option given to `format-pstats` had
keys *almost* but not exactly like the keys in `pstats` maps.

This change makes the two consistent, by fixing the exceptions.

This is an intended to be a NON-BREAKING change: the previous
column names (`:n-calls` and `:total`) will continue to be
supported as aliases.